+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow schema replacement to replace the entire schema
 - Correct histogram anomaly checking
 - Avoid errors with very small `multipleOf` values
+- Avoid errors with `multipleOf` and very different scales
 
 ## [0.15.0]
 ### Changed